+++ src/main/java/kr/bb/product/domain/review/application/port/in/ReviewCommandInputPort.java
@@ -1,10 +1,10 @@
 package kr.bb.product.domain.review.application.port.in;
 
 import java.util.List;
-import kr.bb.product.domain.review.adapter.in.ReviewCommand;
 import kr.bb.product.domain.review.application.port.out.ReviewOutPort;
 import kr.bb.product.domain.review.application.usecase.ReviewCommandUseCase;
 import kr.bb.product.domain.review.entity.Review;
+import kr.bb.product.domain.review.entity.ReviewCommand;
 import kr.bb.product.domain.review.entity.ReviewImages;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
