+++ src/main/java/kr/bb/product/domain/review/application/usecase/ReviewCommandUseCase.java
@@ -1,6 +1,6 @@
 package kr.bb.product.domain.review.application.usecase;
 
-import kr.bb.product.domain.review.adapter.in.ReviewCommand;
+import kr.bb.product.domain.review.entity.ReviewCommand;
 
 public interface ReviewCommandUseCase {
   void writeReview(ReviewCommand.Register review, Long userId, String productId);
